* Cambios por haber hecho el mapo * Comente algunas cosas como para que funcionara y ver dsp como arreglar.

+++ DataLayer/SocioDTO.cs
@@ -7,7 +7,7 @@ namespace DataLayer
     {
 
         //Properties navigation
-        public int DNI { get; set; }
+        public long DNI { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public DateTime FechaNac { get; set; }
